+++ examples/heatmap/BlockHeatMapChart.tsx
@@ -111,7 +111,7 @@ export const BlockHeatMapChart = ({ fref, chartData, rawData }: MilestoneStory) 
                 <Axis variant={'top'}>
                     <AxisTicks
                         variant={'top'}
-                        labelStyle={{ textAnchor: 'start', alignmentBaseline: 'middle' }}
+                        labelStyle={{ textAnchor: 'start', dominantBaseline: 'middle' }}
                     />
                     <AxisLabel variant={'top'}>Replicate 2</AxisLabel>
                 </Axis>