+++ examples/heatmap/CategoricalStripHeatMapChart.tsx
@@ -80,7 +80,7 @@ export const CategoricalStripHeatMapChart = ({ fref, chartData, rawData }: Miles
                     <Axis variant={'top'}>
                         <AxisTicks
                             variant={'top'}
-                            labelStyle={{ textAnchor: 'start', alignmentBaseline: 'middle' }}
+                            labelStyle={{ textAnchor: 'start', dominantBaseline: 'middle' }}
                         />
                         <AxisLabel variant={'top'}>Samples</AxisLabel>
                     </Axis>