+++ examples/heatmap/DualContinuousHeatMapChart.tsx
@@ -95,7 +95,7 @@ export const DualContinuousHeatMapChart = ({ fref, chartData, rawData }: Milesto
                     <Axis variant={'top'}>
                         <AxisTicks
                             variant={'top'}
-                            labelStyle={{ textAnchor: 'start', alignmentBaseline: 'middle' }}
+                            labelStyle={{ textAnchor: 'start', dominantBaseline: 'middle' }}
                         />
                         <AxisLabel variant={'top'}>Samples</AxisLabel>
                     </Axis>