+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 
 ### Changed
 
+-   [packages/themes]: changes styles in some theme pieces to be more compatible with inkscape svg
 -   [packages/core]: changed implementation of `Typography` and `Label` to allow animations of rotated text
 -   [packages/annotation]: improved `cleanSvg` to simplify scaleX, scaleY, and rotate transformations
 -   [packages/annotation]: changed encoding of svg into text files to include newline characters